+++ src/components/2017/page-book-review-more-than-everything-by-vanessa-foster/page-book-review-more-than-everything-by-vanessa-foster.tsx
@@ -1,5 +1,5 @@
 import { Component, Element, h } from '@stencil/core';
-import { extractIdFromDocumentPath, SITENAME } from '../../../helpers/utils';
+import { extractIdFromDocumentPath } from '../../../helpers/utils';
 import { BlogData } from '../../../services/blog-data';
 
 @Component({
@@ -14,12 +14,6 @@ export class PageBookReviewMoreThanEverythingByVanessaFoster {
   componentWillLoad() {
     let id = extractIdFromDocumentPath();
     this.header = BlogData.getPostHeaderById(id);
-
-    // set document title for browser / tab / bookmark
-    document.title = this.header.title + ' | ' + SITENAME;
-    if (this.header.teaser) {
-      document.getElementById("meta-desc").setAttribute("content", this.header.teaser);
-    }
   }
 
   toggleSearch() {