+++ src/components/2017/page-sparql-examples-construct/page-sparql-examples-construct.tsx
@@ -1,5 +1,5 @@
 import { Component, h } from '@stencil/core';
-import { extractIdFromDocumentPath, SITENAME } from '../../../helpers/utils';
+import { extractIdFromDocumentPath } from '../../../helpers/utils';
 import { BlogData } from '../../../services/blog-data';
 import '@deckdeckgo/highlight-code';
 
@@ -13,12 +13,6 @@ export class PageSparqlExamplesConstruct {
 	async componentWillLoad() {
 		let id = extractIdFromDocumentPath();
 		this.header = BlogData.getPostHeaderById(id);
-
-		// set document title for browser / tab / bookmark
-		document.title = this.header.title + ' | ' + SITENAME;
-		if (this.header.teaser) {
-			document.getElementById("meta-desc").setAttribute("content", this.header.teaser);
-		}
 	}
 
 	render() {