+++ src/components/2017/page-use-spring-for-stardog/page-use-spring-for-stardog.tsx
@@ -1,5 +1,5 @@
 import { Component, h } from '@stencil/core';
-import { extractIdFromDocumentPath, SITENAME } from '../../../helpers/utils';
+import { extractIdFromDocumentPath } from '../../../helpers/utils';
 import { BlogData } from '../../../services/blog-data';
 import '@deckdeckgo/highlight-code';
 
@@ -13,11 +13,6 @@ export class PageUseSpringForStardog {
     async componentWillLoad() {
         let id = extractIdFromDocumentPath();
         this.header = BlogData.getPostHeaderById(id);
-
-        document.title = this.header.title + ' | ' + SITENAME;
-        if (this.header.teaser) {
-            document.getElementById("meta-desc").setAttribute("content", this.header.teaser);
-        }
     }
 
     render() {