+++ src/components/2018/page-beaver-cage-command-chron/page-beaver-cage-command-chron.tsx
@@ -1,5 +1,5 @@
 import { Component, Element, h } from '@stencil/core';
-import { extractIdFromDocumentPath, SITENAME } from '../../../helpers/utils';
+import { extractIdFromDocumentPath} from '../../../helpers/utils';
 import { BlogData } from '../../../services/blog-data';
 
 @Component({
@@ -15,11 +15,6 @@ export class PageBeaverCageCommandChron {
     async componentWillLoad() {
         let id = extractIdFromDocumentPath();
         this.header = BlogData.getPostHeaderById(id);
-
-        document.title = this.header.title + ' | ' + SITENAME;
-        if (this.header.teaser) {
-            document.getElementById("meta-desc").setAttribute("content", this.header.teaser);
-        }
     }
 
     toggleSearch() {