+++ src/components/2018/page-cage/page-cage.tsx
@@ -1,5 +1,5 @@
 import { Component, Element, h } from '@stencil/core';
-import { extractIdFromDocumentPath, SITENAME } from '../../../helpers/utils';
+import { extractIdFromDocumentPath } from '../../../helpers/utils';
 import { BlogData } from '../../../services/blog-data';
 
 @Component({
@@ -21,17 +21,10 @@ export class PageCage {
     let id = extractIdFromDocumentPath();
     this.header = BlogData.getPostHeaderById(id);
 
-    // set document title for browser / tab / bookmark
-    document.title = this.header.title + ' | ' + SITENAME;
-    if (this.header.teaser) {
-        document.getElementById("meta-desc").setAttribute("content", this.header.teaser);
-    }
-
     this.documentItems = await BlogData.getPostsByMenu('cage-documents');
     this.photoItems = await BlogData.getPostsByMenu('cage-photos');
     this.videoItems = await BlogData.getPostsByMenu('cage-videos'); 
-    this.referenceItems = await BlogData.getPostsByMenu('cage-ref-resources');
-    
+    this.referenceItems = await BlogData.getPostsByMenu('cage-ref-resources'); 
 }
 
   toggleSearch(){