+++ src/components/2020/page-sea-tiger-newspapers/page-sea-tiger-newspapers.tsx
@@ -1,5 +1,5 @@
 import { Component, Element, h } from '@stencil/core';
-import { extractIdFromDocumentPath, SITENAME } from '../../../helpers/utils';
+import { extractIdFromDocumentPath } from '../../../helpers/utils';
 import { BlogData } from '../../../services/blog-data';
 
 @Component({
@@ -158,11 +158,6 @@ export class PageSeaTigerNewspapers {
   async componentWillLoad() {
     let id = extractIdFromDocumentPath();
     this.header = BlogData.getPostHeaderById(id);
-
-    document.title = this.header.title + ' | ' + SITENAME;
-    if (this.header.teaser) {
-      document.getElementById("meta-desc").setAttribute("content", this.header.teaser);
-    }
   }
 
   toggleSearch() {