+++ changelog.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is loosely based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.9.0 - May 21, 2020 - PRERENDERED]
+
+- Update page: [/cage/vietnam-war-reference-resources](http://codyburleson.com/cage/vietnam-war-reference-resources); add reference to video documentary series on Tubi, US Marines in Vietnam.
+- Update page: [/lifetime](https://codyburleson.com/lifetime); remove side column, allowing calendar to widen up on the screen.
+- Refactor so that the setting of page titles and meta tags are done all in one single place (in the app-root.tsx file). Before, ponly the pages were having titles set there and then the posts were doing it in each post's component. Now, this not only puts the code in one single place, but it seems to be the only place where it works effectively also when prerendering (and I am trying to head into a fully prerendered site, eventually).
+- Got prerendering working for at least the main pages and then, even when prerendered, the rest works as it normally did. So, this marks the first deployment of prerendered site!
+
 ## [1.8.2 - May 20, 2020]
 
 - Fixed broken page: [/introduction-to-the-semantic-web-vision-and-technologies](https://codyburleson.com/introduction-to-the-semantic-web-vision-and-technologies); uploaded video version of presentation to YouTube and embedded it; uploaded zipped PowerPoint to Amazon S3 and linked to it.